Added CWD function for Linux compatibility

+++ candidateprocessor.py
@@ -15,7 +15,11 @@ from PIL import Image, ImageDraw, ImageFont
 from imgur_downloader import ImgurDownloader
 from moviepy.editor import VideoFileClip, concatenate_videoclips
 
-CWD = os.getcwd()+"/"
+
+def CWD(dir):
+    """ Helper function to join the current working directory to another path
+        Used for Linux compatibility """
+    return os.path.join(os.getcwd(), dir)
 
 def generate_sign(name, position, district, party, image, output):
     sign_templates = {
@@ -29,14 +33,11 @@ def generate_sign(name, position, district, party, image, output):
     img = img.resize((2197, 3304))
     sign.paste(img, (2906, 0))
 
-    font = ImageFont.truetype(CWD+"Assets/font.ttf", 250)
-    font_name = ImageFont.truetype(CWD+"Assets/font.ttf", 300)
+    font = ImageFont.truetype(CWD("Assets/font.ttf"), 300)
     draw = ImageDraw.Draw(sign)
-    text = f"{position}\nOf\n{district}"
-    name_w, name_h = draw.textsize(text, font=font_name)
+    text = f"{name}\n{position}\nOf\n{district}"
     text_w, text_h = draw.multiline_textsize(text, font=font, spacing=20)
-    draw.text(((2401 - name_w) / 2 + 253, 1400), name, (0, 0, 0), font=font_name, align="center")
-    draw.multiline_text(((2401 - text_w) / 2 + 253, 1750), text, (0, 0, 0),
+    draw.multiline_text(((2401 - text_w) / 2 + 253, 1400), text, (0, 0, 0),
                         font=font, align="center", spacing=20)
 
     sign.save(output, quality=95)
@@ -54,23 +55,23 @@ def generate_video(sign, issue, output):
         "Education Funding": "EducationFunding.mp4"
     }
 
-    video_path = CWD+f"Assets/{videos[issue]}"
+    video_path = CWD(f"Assets/{videos[issue]}")
 
     frame = cv2.imread(sign)
     frame = cv2.resize(frame, (1920, 1080))
     height, width, layers = frame.shape
     fourcc = cv2.VideoWriter_fourcc(*'mp4v')
-    video = cv2.VideoWriter(CWD+"temp.mp4", fourcc, 1, (width, height))
+    video = cv2.VideoWriter(CWD("temp.mp4"), fourcc, 1, (width, height))
     for i in range(5):
         video.write(frame)
     video.release()
 
-    image_clip = VideoFileClip(CWD+"temp.mp4")
+    image_clip = VideoFileClip(CWD("temp.mp4"))
     original_video = VideoFileClip(video_path)
     final_video = concatenate_videoclips([original_video, image_clip], method="compose")
 
     final_video.write_videofile(output)
-    os.remove(CWD+"temp.mp4")
+    os.remove(CWD("temp.mp4"))
 
 
 def send_email(receiver_email, filename):
@@ -80,7 +81,7 @@ def send_email(receiver_email, filename):
     body = "Thank you for using the Campaign Generator! Submit another campaign here: https://forms.gle/iTE4DXoEEpYytyru8"
     sender_email = "campaigngeneratorjwu@gmail.com"
 
-    with open(CWD+"mailcreds.txt", "r") as file:
+    with open(CWD(os.path.join("..", "mailcreds.txt")), "r") as file:
         password = file.read()
 
     # Create a multipart message and set headers
@@ -119,6 +120,32 @@ def send_email(receiver_email, filename):
         server.login(sender_email, password)
         server.sendmail(sender_email, receiver_email, text)
 
+def send_error_email(receiver_email, subject, body):
+    """ Sends an email, used for errors """
+
+    sender_email = "campaigngeneratorjwu@gmail.com"
+
+    with open(CWD("mailcreds.txt"), "r") as file:
+        password = file.read()
+
+    # Create a multipart message and set headers
+    message = MIMEMultipart()
+    message["From"] = sender_email
+    message["To"] = receiver_email
+    message["Subject"] = subject
+    message["Bcc"] = receiver_email  # Recommended for mass emails
+
+    # Add body to email
+    message.attach(MIMEText(body, "plain"))
+
+    text = message.as_string()
+
+    # Log in to server using secure context and send email
+    context = ssl.create_default_context()
+    with smtplib.SMTP_SSL("smtp.gmail.com", 465, context=context) as server:
+        server.login(sender_email, password)
+        server.sendmail(sender_email, receiver_email, text)
+
 
 def process_candidate(candidate_list, send_to_email=True):
     """ Main function that generates the sign, video, and speech using the helper functions,
@@ -128,17 +155,27 @@ def process_candidate(candidate_list, send_to_email=True):
     for current_candidate in candidate_list:
         # Create the campaign sign
         # Get the candidate's image from Imgur
-        ImgurDownloader(current_candidate["image"]).save_images(CWD+"candidate_images")  # Download Imgur image/album
-        candidate_image = CWD+f"candidate_images/{os.listdir('candidate_images')[0]}"  # Find the first image in the album
+        try:
+            ImgurDownloader(current_candidate["image"]).save_images(CWD("candidate_images"))  # Download Imgur image/album
+        except:
+            print("Error getting Imgur link, send error email...")
+            send_error_email(current_candidate["email"],
+                             "Campaign Generator Failed",
+                             "The campaign generator failed to get your image from Imgur, please make sure you are using the direct link to the image (i.imgur.com)\n"+
+                             "Resubmit your campaign here: https://forms.gle/mAwfNwK9SSLR1Riy9")
+            return
+        candidate_image = CWD(f"candidate_images/{os.listdir('candidate_images')[0]}")  # Find the first image in the album
 
         # Generate the sign with the candidates data
-        sign_output = CWD+f"output/{current_candidate['name']} sign.png"
+        print(f"Generating campaign sign for {current_candidate['name']}")
+        sign_output = CWD(f"output/{current_candidate['name']} sign.png")
         generate_sign(current_candidate["name"], current_candidate["position"],
                       current_candidate["district"], current_candidate["party"],
                       candidate_image, sign_output)
-        rmtree(CWD+"candidate_images")  # Delete the images folder
+        rmtree(CWD("candidate_images"))  # Delete the images folder
 
         # Create the campaign speech
+
         speech = f"""My Fellow Americans,
 
 I stand before you today not to spew fancy political rhetoric or to talk about what could be better. I’m here today because you all are here. And you’re clearly fed up with politics. I’m here because our democracy isn’t working for all of us. You all know that this is the most important election of our lifetime. So I’m going to do something. This is why I, {current_candidate['name']}, am running to be the {current_candidate['position']} of {current_candidate['district']}.
@@ -150,19 +187,19 @@ However, putting our all towards {current_candidate['issues'][0]} won’t be eno
 Vote {current_candidate['name']} for {current_candidate['position']} of {current_candidate['district']}!
 """
         # Write the speech to a file
-        with open(CWD+f"output/{current_candidate['name']} speech.txt", "w") as file:
+        with open(CWD(f"output/{current_candidate['name']} speech.txt"), "w") as file:
             file.write(speech)
 
         generate_video(sign_output, current_candidate["issues"][0],
-                       CWD+f"output/{current_candidate['name']} video.mp4")
+                       CWD(f"output/{current_candidate['name']} video.mp4"))
 
         # Zip all files in output folder
-        file_paths = os.listdir(CWD+"output")
-        with ZipFile(CWD+'output\\Generated Campaign.zip', 'w') as zipped:
+        file_paths = os.listdir(CWD("output"))
+        with ZipFile(os.path.join(CWD('output'), 'Generated Campaign.zip'), 'w') as zipped:
             # writing each file one by one
-            os.chdir(CWD+"\\output")
+            os.chdir(CWD("output"))
             for file in file_paths:
-                if "zip" not in file:
+                if "zip" not in file and current_candidate["name"] in file:
                     zipped.write(file, arcname=file)
                     os.remove(file)
             zipped.close()
